+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/option/SoraChannelRole.kt
@@ -6,13 +6,14 @@ import java.util.*
  * チャネルの役割を示します
  */
 enum class SoraChannelRole {
-    /** 配信 */
-    UPSTREAM,
-    /** 視聴 */
-    DOWNSTREAM,
 
+    /** 送信のみ */
     SENDONLY,
+
+    /** 受信のみ */
     RECVONLY,
+
+    /** 送受信 */
     SENDRECV;
 
     internal val signaling: String