+++ CHANGES.md
@@ -19,6 +19,10 @@
   - @szktty
 - `com.squareup.okhttp3:okhttp` を 4.7.2 に上げる
   - @szktty
+- 新しいロール (`sendonly`, `recvonly`, `sendrecv`) に対応する
+  - @szktty
+- 古いロール (`upstream`, `downstream`) を削除する
+  - @szktty
 - `SoraAudioOption.audioSource` のデフォルト値を `VOICE_COMMUNICATION` に変更する
   - @szktty